Store validator assignments in memory to optimize fetches

+++ beacon-chain/rpc/validator_server.go
@@ -131,9 +131,7 @@ func (vs *ValidatorServer) CommitteeAssignment(
 	if err != nil {
 		return nil, fmt.Errorf("could not get active validator index: %v", err)
 	}
-
-	committee, shard, slot, isProposer, err :=
-		helpers.CommitteeAssignment(beaconState, req.EpochStart, uint64(idx), false)
+	committee, shard, slot, isProposer, err := helpers.CommitteeAssignment(beaconState, req.EpochStart, idx, false)
 	if err != nil {
 		return nil, err
 	}
@@ -141,6 +139,7 @@ func (vs *ValidatorServer) CommitteeAssignment(
 	if err != nil {
 		return nil, err
 	}
+
 	return &pb.CommitteeAssignmentResponse{Assignment: []*pb.CommitteeAssignmentResponse_CommitteeAssignment{
 		{
 			Committee:  committee,
